Linux: remove 2 URB transmit path
Since a previous commit ensured we can always fit a CAN(-FD) message
into the CAN msg buffer, this path is now obsolete.

For reference, the path would use two TX URBs to send a CAN-FD message
with a DLC>=14.

Frankly I don't remember why this path was introduced in the first place.
I recon it must have been during the time I discovered that Windows
jumbles the order of the RX URBs when signaling completion.