Windows: DLL: adds support for external rx wait loop
A user may want to wait on more than one handle in addition to
the USB wait handle. Instead of letting the user pass in any number
of handles, enable retrieval of the next USB wait handle.

This way the user is in control, the library avoids handle copying
and the library internal logic is much simpler.

The API call is sc_can_stream_rx_next_wait_handle(stream, [out] handle)